idempotent global tracer for testing, and capturing some basic traces from the tests directly
