+++ Packages/automerge-repo/Sources/AutomergeRepo/Networking/NetworkProvider.swift
@@ -1,4 +1,4 @@
-import AsyncAlgorithms
+// import AsyncAlgorithms
 
 // import protocol Combine.Publisher
 import Automerge
