+++ Packages/automerge-repo/Sources/AutomergeRepo/Networking/NetworkSubsystem.swift
@@ -1,4 +1,4 @@
-import AsyncAlgorithms
+// import AsyncAlgorithms
 import Automerge
 import struct Foundation.Data
 import OSLog
