+++ Packages/automerge-repo/notes.md
@@ -1,0 +1,10 @@
+# using docker-compose
+
+`docker-compose -f someDockerComposefile up -d`, for example:
+
+```bash
+docker-compose -f docker-compose.yml up -d
+```
+
+there's an equiv for Tempo, and another for sigNoz
+https://github.com/SigNoz/signoz/tree/develop/deploy/docker/clickhouse-setup
